+++ src/just_makeit/templates/cmake/CMakeLists_object_core.cmake
@@ -1,7 +1,8 @@
 <<object_core_decl>>
 <<extra_include_dirs_on_object_core>><<extra_link_on_object_core>>
 add_executable(test_<<component>>_core
-               ${CMAKE_SOURCE_DIR}/native/tests/test_<<component>>_core.c)
+               ${CMAKE_SOURCE_DIR}/native/tests/test_<<component>>_core.c
+               ${CMAKE_SOURCE_DIR}/native/tests/test_<<component>>_symbols.c)
 target_link_libraries(test_<<component>>_core
                       PRIVATE <<component>>_core <<extra_link_libs_block>>${JM_MATH_LIBRARY})
 target_include_directories(test_<<component>>_core